+++ src/components/MenuBar.js
@@ -1,5 +1,8 @@
 
-
+import {
+  signOut,
+  useSession
+} from "next-auth/client"
 
 import styles from "../styles/MenuBar.module.css";
 
@@ -10,10 +13,10 @@ import MenuItem from "../components/MenuItem";
 
 
 
+
 export default function MenuBar({allHobbies, children}){
   const [visible, toggleMenu] = useState(false);
-
-
+  const [ session ] = useSession()
 
   const icon = (
     <span onClick={() => {toggleMenu(!visible)}} className = {styles.menuHeaderI}>
@@ -24,7 +27,8 @@ export default function MenuBar({allHobbies, children}){
     <div className={styles.container}>
     &emsp;&emsp;Hobby Buddy
     <br/>
-    <button className = {styles.button1}>+Add Event</button>
+
+    <button className = {styles.button1} onClick={signOut}>Sign Out</button>
     <ul className={styles.ul}>
 
       <MenuItem title = {"Profile"} icon = {"fa fa-user-circle"}/>
