+++ src/pages/_app.js
@@ -1,8 +1,13 @@
 /* eslint-disable */
 import '../styles/globals.css';
+import { Provider } from 'next-auth/client';
 
 function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />;
+ return (
+  <Provider session={pageProps.session}>
+    <Component {...pageProps} />
+  </Provider>
+);
 }
 
 export default MyApp;
